@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFrameworks>netcoreapp3.1;net5.0</TargetFrameworks>
-        <LangVersion>9</LangVersion>
+        <TargetFrameworks>netcoreapp3.1;net6.0</TargetFrameworks>
+        <LangVersion>10</LangVersion>
         <IsPackable>false</IsPackable>
         <Nullable>enable</Nullable>
     </PropertyGroup>
@@ -20,15 +20,15 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Crip.AspNetCore.Logging\Crip.AspNetCore.Logging.csproj" />
-        <ProjectReference Include="..\Crip.AspNetCore.Tests\Crip.AspNetCore.Tests.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
         <None Remove="Resources\sis_vehicle_response.json" />
         <EmbeddedResource Include="Resources\sis_vehicle_response.json" />
         <None Remove="Resources\sis_vehicle_response_log.json" />
         <EmbeddedResource Include="Resources\sis_vehicle_response_log.json" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Crip.AspNetCore.Logging\Crip.AspNetCore.Logging.csproj" />
+      <ProjectReference Include="..\Crip.AspNetCore.Tests\Crip.AspNetCore.Tests.csproj" />
     </ItemGroup>
 
 </Project>
