@@ -1,47 +1,53 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <TargetFrameworks>netcoreapp3.1;net5.0</TargetFrameworks>
-        <LangVersion>9</LangVersion>
-        <Nullable>enable</Nullable>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <DocumentationFile>$(AssemblyName).xml</DocumentationFile>
-        <PackageId>$(AssemblyName)</PackageId>
-        <Version>1.0.0.0</Version>
-        <Authors>TAHQ69</Authors>
-        <Company>CRIP</Company>
-        <Title>$(AssemblyName)</Title>
-        <Description>AspNetCore HTTP request logging.</Description>
-        <PackageDescription>AspNetCore request logging.</PackageDescription>
-        <RepositoryUrl>https://github.com/crip-home/Crip.AspNetCore.Logging</RepositoryUrl>
-        <PackageProjectUrl>https://github.com/crip-home/Crip.AspNetCore.Logging</PackageProjectUrl>
-        <PackageLicenseUrl>https://github.com/crip-home/Crip.AspNetCore.Logging/blob/master/LICENSE</PackageLicenseUrl>
-        <RepositoryType>git</RepositoryType>
-        <PackageTags>AspNetCore, HTTP, log</PackageTags>
-    </PropertyGroup>
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <LangVersion>10.0</LangVersion>
+    <Nullable>enable</Nullable>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <DocumentationFile>$(AssemblyName).xml</DocumentationFile>
+    <PackageId>$(AssemblyName)</PackageId>
+    <Version>1.0.0.0</Version>
+    <Authors>TAHQ69</Authors>
+    <Company>CRIP</Company>
+    <Title>$(AssemblyName)</Title>
+    <Description>AspNetCore HTTP request logging.</Description>
+    <PackageDescription>AspNetCore request logging.</PackageDescription>
+    <RepositoryUrl>https://github.com/crip-home/Crip.AspNetCore.Logging</RepositoryUrl>
+    <PackageProjectUrl>https://github.com/crip-home/Crip.AspNetCore.Logging</PackageProjectUrl>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <PackageReadmeFile>README.MD</PackageReadmeFile>
+    <RepositoryType>git</RepositoryType>
+    <PackageTags>AspNetCore, HTTP, log</PackageTags>
+  </PropertyGroup>
 
-    <PropertyGroup>
-      <NoWarn>1701;1702;SA1101;SA1633;SA1200;SA1309</NoWarn>
-    </PropertyGroup>
+  <PropertyGroup>
+    <NoWarn>1701;1702;SA1101;SA1633;SA1200;SA1309;SA1501;SA1503</NoWarn>
+  </PropertyGroup>
 
-    <ItemGroup>
-        <FrameworkReference Include="Microsoft.AspNetCore.App" />
-    </ItemGroup>
+  <ItemGroup>
+    <None Remove="$(AssemblyName).xml" />
+    <None Include="../README.MD" Pack="true" PackagePath="\" />
+    <None Include="../LICENSE" Pack="true" PackagePath="\" />
+  </ItemGroup>
 
-    <ItemGroup>
-        <None Remove="$(AssemblyName).xml" />
-    </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="$(AssemblyName).xml" />
+  </ItemGroup>
 
-    <ItemGroup>
-        <EmbeddedResource Include="$(AssemblyName).xml" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
-        <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.261">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-    </ItemGroup>
+  <ItemGroup>
+    
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
+    
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
+    
+    <PackageReference Include="Microsoft.Extensions.Http" Version="2.2.0" />
+    
+    <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.261">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+  </ItemGroup>
 
 </Project>
