@@ -1,11 +1,10 @@
 ﻿using System.Net.Http;
 
-namespace Crip.AspNetCore.Logging
+namespace Crip.AspNetCore.Logging;
+
+/// <summary>
+/// Non generic logging handler abstraction.
+/// </summary>
+public abstract class LoggingHandler : DelegatingHandler
 {
-    /// <summary>
-    /// Non generic logging handler abstraction.
-    /// </summary>
-    public abstract class LoggingHandler : DelegatingHandler
-    {
-    }
 }
