@@ -1,20 +1,19 @@
-namespace Crip.AspNetCore.Logging
+namespace Crip.AspNetCore.Logging;
+
+/// <summary>
+/// Time measurement service contract.
+/// </summary>
+public interface IMeasurable
 {
     /// <summary>
-    /// Time measurement service contract.
+    /// Start measure time.
     /// </summary>
-    public interface IMeasurable
-    {
-        /// <summary>
-        /// Start measure time.
-        /// </summary>
-        /// <returns>Started time measurement service.</returns>
-        IMeasurable StartMeasure();
+    /// <returns>Started time measurement service.</returns>
+    IMeasurable StartMeasure();
 
-        /// <summary>
-        /// Stop time measure and get measurement result.
-        /// </summary>
-        /// <returns>Stopped time measurement state.</returns>
-        IStopwatch StopMeasure();
-    }
+    /// <summary>
+    /// Stop time measure and get measurement result.
+    /// </summary>
+    /// <returns>Stopped time measurement state.</returns>
+    IStopwatch StopMeasure();
 }
