@@ -1,11 +1,10 @@
-﻿namespace System.Runtime.CompilerServices
+﻿namespace System.Runtime.CompilerServices;
+
+/// <summary>
+/// Using `init` feature in c# 9 raise error: Predefined type
+/// 'System.Runtime.CompilerServices.IsExternalInit' is not defined or
+/// imported.
+/// </summary>
+public sealed class IsExternalInit
 {
-    /// <summary>
-    /// Using `init` feature in c# 9 raise error: Predefined type
-    /// 'System.Runtime.CompilerServices.IsExternalInit' is not defined or
-    /// imported.
-    /// </summary>
-    public sealed class IsExternalInit
-    {
-    }
 }
