@@ -1,7 +1,7 @@
 using System.Net.Mime;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Crip.AspNetCore.Logging.Example.Web.Controllers
+namespace Crip.AspNetCore.Logging.Core31.Example.Controllers
 {
     [Produces(MediaTypeNames.Application.Json)]
     [ApiController]
